BID/JTax/publish/requirements.txt
<<<<<<< .mine
amqp==2.1.4

appdirs==1.4.3

billiard==3.5.0.2

celery==4.0.2

Django==1.11
djangorestframework == 3.7.7
django-filter
django-cors-headers==2.1.0

django-haystack==2.6.0

jieba==0.38

kombu==4.0.2

packaging==16.8
#pkg-resources==0.0.0

psycopg2==2.7.3.2

pymongo==3.4.0

PyMySQL==0.7.10

pyparsing==2.2.0

pytz==2016.10

#redis==2.10.5

requests==2.18.4

six==1.10.0

vine==1.1.3

Whoosh==2.7.4

category==0.9.10

simplejson == 3.13.2

qcloudsms-py==0.1.1
||||||| .r1029
amqp==2.1.4

appdirs==1.4.3

billiard==3.5.0.2

celery==4.0.2

Django==1.11
djangorestframework == 3.7.7
django-filter
django-cors-headers==2.1.0

django-haystack==2.6.0

jieba==0.38

kombu==4.0.2

packaging==16.8
#pkg-resources==0.0.0

psycopg2==2.7.3.2

pymongo==3.4.0

PyMySQL==0.7.10

pyparsing==2.2.0

pytz==2016.10

#redis==2.10.5

requests==2.18.4

six==1.10.0

vine==1.1.3

Whoosh==2.7.4

category==0.9.10

simplejson == 3.13.2

qcloudsms-py
=======
amqp==2.1.4
appdirs==1.4.3
billiard==3.5.0.2
category==0.9.10
celery==4.0.2
certifi==2018.1.18
chardet==3.0.4
class-registry==2.1.2
coreapi==2.1.1
coreschema==0.0.4
Django==1.11
django-appconf==1.0.2
django-compressor==1.5
django-cors-headers==2.1.0
django-crispy-forms==1.7.2
django-filter==1.1.0
django-haystack==2.6.0
djangorestframework==3.8.2
idna==2.6
itypes==1.1.0
jieba==0.38
Jinja2==2.10
kombu==4.0.2
Markdown==2.6.2
MarkupSafe==1.0
packaging==16.8
psycopg2==2.7.3.2
pyaml==16.12.2
pymongo==3.4.0
PyMySQL==0.7.10
pyparsing==2.2.0
python-dateutil==2.7.2
pytz==2016.10
PyYAML==3.12
qcloudsms-py==0.1.1
redis==2.10.5
regex==2018.2.21
requests==2.18.4
simplejson==3.13.2
six==1.10.0
uritemplate==3.0.0
urllib3==1.22
vine==1.1.3
Whoosh==2.7.4
>>>>>>> .r1059
